generator/requirements.txt
wheel
google-cloud-pubsub==0.42.1
StringGenerator==0.3.3
avro-python3==1.9.1
